create discid package / split module, re #25
This splits the discid module into multiple modules:

libdiscid.py loads libdiscid and extracts the constants
disc.py is the Disc class and related functions (read/put)
util.py has utility functsions (_encode/_decode)
